cache refactors

+++ VarCache.cpp
@@ -88,7 +88,7 @@ namespace jasl {
         return OptionalValueArray();
     }
 
-    Value VarCache::getListToken(std::string const &key, int const index)
+    Value VarCache::getListToken(std::string const &key, size_t const index)
     {
         auto it = listCache.find(key);
         if(it != std::end(listCache))  {
